🏗️ Refactor(character): Move all dtos and schemas to a core folder for better reading in the character module
