+++ backend/src/resources/character/core/constants/sizes.constant.ts
