+++ backend/src/resources/character/core/dto/ability/ability.dto.ts
