+++ backend/src/resources/character/core/dto/affinities/affinities.dto.ts
