+++ backend/src/resources/character/core/dto/spellcasting/sub/spell.dto.ts
