+++ backend/src/resources/character/core/dto/stats/sub/skill.dto.ts
