+++ backend/src/resources/character/core/dto/stats/sub/speed.dto.ts
