+++ backend/src/resources/character/core/dto/update-character.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateCharacterDto } from '@/resources/character/dto/create-character.dto';
+import { CreateCharacterDto } from '@/resources/character/core/dto/create-character.dto';
 
 export class UpdateCharacterDto extends PartialType(CreateCharacterDto) {}
