+++ backend/src/resources/character/core/schemas/ability/ability.schema.ts
