+++ backend/src/resources/character/core/schemas/affinities/affinities.schema.ts
