+++ backend/src/resources/character/core/schemas/character.schema.ts
@@ -1,11 +1,11 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import mongoose, { Document } from 'mongoose';
-import { Affinities } from '@/resources/character/schemas/affinities/affinities.schema';
+import { Affinities } from '@/resources/character/core/schemas/affinities/affinities.schema';
 import { Group } from '@/resources/group/schemas/group.schema';
 import { BaseSchema } from '@/common/schemas/base-schema';
-import { Ability } from '@/resources/character/schemas/ability/ability.schema';
-import { Spellcasting } from '@/resources/character/schemas/spellcasting/spellcasting.schema';
-import { Stats } from '@/resources/character/schemas/stats/stats.schema';
+import { Ability } from '@/resources/character/core/schemas/ability/ability.schema';
+import { Spellcasting } from '@/resources/character/core/schemas/spellcasting/spellcasting.schema';
+import { Stats } from '@/resources/character/core/schemas/stats/stats.schema';
 
 export type CharacterDocument = Character & Document;
