+++ backend/src/resources/character/core/schemas/profile/profile.schema.ts
