+++ backend/src/resources/character/core/schemas/spellcasting/sub/spell.schema.ts
