+++ backend/src/resources/character/core/schemas/stats/sub/abilityScores.schema.ts
