+++ backend/src/resources/character/core/schemas/stats/sub/skill.schema.ts
