+++ backend/src/resources/character/core/schemas/stats/sub/speed.schema.ts
