+++ backend/src/resources/character/npc/dto/create-npc.dto.ts
@@ -1,9 +1,9 @@
 import { Type } from 'class-transformer';
 import { IsArray, IsOptional, ValidateNested } from 'class-validator';
-import { ProfileDto } from '@/resources/character/dto/profile/profile.dto';
+import { ProfileDto } from '@/resources/character/core/dto/profile/profile.dto';
 import { ActionsDto } from '@/resources/character/npc/dto/actions/actions.dto';
 import { ChallengeDto } from '@/resources/character/npc/dto/challenge/challenge.dto';
-import { CreateCharacterDto } from '@/resources/character/dto/create-character.dto';
+import { CreateCharacterDto } from '@/resources/character/core/dto/create-character.dto';
 
 export class CreateNpcDto extends CreateCharacterDto {
   @ValidateNested({ each: true })
