+++ backend/src/resources/character/npc/schemas/npc.schema.ts
@@ -3,7 +3,7 @@ import { Document } from 'mongoose';
 import { Actions } from '@/resources/character/npc/schemas/actions/actions.schema';
 import { Challenge } from '@/resources/character/npc/schemas/challenge/challenge.schema';
 import { NPCProfile } from '@/resources/character/npc/schemas/profile/npc-profile.schema';
-import { Character } from '@/resources/character/schemas/character.schema';
+import { Character } from '@/resources/character/core/schemas/character.schema';
 
 export type NPCDocument = NPC & Document;
