+++ backend/src/resources/character/player/dto/profile/player-profile.dto.ts
@@ -1,5 +1,5 @@
 import { IsOptional, IsString } from 'class-validator';
-import { ProfileDto } from '@/resources/character/dto/profile/profile.dto';
+import { ProfileDto } from '@/resources/character/core/dto/profile/profile.dto';
 
 export class PlayerProfileDto extends ProfileDto {
   @IsOptional()
