+++ backend/src/resources/character/player/dto/stats/player-stats.dto.ts
@@ -1,5 +1,5 @@
 import { IsNumber, IsOptional } from 'class-validator';
-import { StatsDto } from '@/resources/character/dto/stats/stats.dto';
+import { StatsDto } from '@/resources/character/core/dto/stats/stats.dto';
 
 export class PlayerStatsDto extends StatsDto {
   @IsOptional()
